Changed time display for results. Enabled bolding of query in search results
