+++ src/js/content.js
@@ -1,44 +1,5 @@
 import React from 'react';
-
-const timeSince = (date) => {
-  const seconds = Math.floor((new Date() - date) / 1000);
-  let interval = Math.floor(seconds / 31536000);
-  if (interval > 1) {
-    return `${interval} years ago`;
-  }
-  if (interval === 1) {
-    return `${interval} year ago`;
-  }
-  interval = Math.floor(seconds / 2592000);
-  if (interval > 1) {
-    return `${interval} months ago`;
-  }
-  if (interval === 1) {
-    return `${interval} month ago`;
-  }
-  interval = Math.floor(seconds / 86400);
-  if (interval > 1) {
-    return `${interval} days ago`;
-  }
-  if (interval === 1) {
-    return `${interval} day ago`;
-  }
-  interval = Math.floor(seconds / 3600);
-  if (interval > 1) {
-    return `${interval} hours ago`;
-  }
-  if (interval === 1) {
-    return `${interval} hour ago`;
-  }
-  interval = Math.floor(seconds / 60);
-  if (interval > 1) {
-    return `${interval} minutes ago`;
-  }
-  if (interval === 1) {
-    return `${interval} minute ago`;
-  }
-  return `${Math.floor(seconds)} seconds ago`;
-};
+import timeSince from './timeSince';
 
 const Content = ({ result }) => (
   <div className="content">
